requirements.txt
flask
pdfplumber
python-docx
pytesseract
pillow
openai
python-dotenv
